rename trainer_hooks -> trainer
