+++ examples/mpt_continued_training.py
@@ -23,7 +23,7 @@ from levanter.logging import capture_time, log_time_to_wandb
 from levanter.models.gpt2 import Gpt2LMHeadModel
 from levanter.models.loss import next_token_loss
 from levanter.models.mpt import MptLmHeadModel
-from levanter.trainer_hooks import StepInfo, TrainerHooks
+from levanter.trainer import StepInfo, TrainerHooks
 from levanter.utils.jax_utils import global_key_array, parameter_count
 from levanter.utils.py_utils import non_caching_cycle
