+++ src/levanter/callbacks.py
@@ -22,7 +22,7 @@ import levanter.visualization as viz
 import wandb
 from levanter.config import WandbConfig
 from levanter.logging import log_optimizer_hyperparams, save_xla_dumps_to_wandb
-from levanter.trainer_hooks import StepInfo
+from levanter.trainer import StepInfo
 
 
 logger = logging.getLogger(__name__)
